requirements.txt
# Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.5

# Database
# Para PostgreSQL em produção, descomente a linha abaixo
# psycopg2-binary==2.9.9

# HTTP & APIs
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.3

# Visualization
plotly==5.18.0

# Utilities
python-decouple==3.8
pycountry==23.12.11
pytz==2023.3
Pillow==10.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Security
python-dotenv==1.0.0
cryptography==41.0.7

# Monitoring & Logging (Optional)
# sentry-sdk==1.38.0

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1
